Separate logging-service and logger npm package
trubudget-pino-logger - npm package
logging-service - trubudget component
Fixed wrong error handling when verifying api token in logging service
