+++ logging-service/.env_example
