+++ logging-service/Dockerfile
@@ -1,14 +1,13 @@
 FROM node:18.4.0-alpine
 
 WORKDIR /home/node
-COPY index.js index.js
-COPY serve.js serve.js
-COPY package.json ./
-COPY package-lock.json ./
+# TODO Remove this after publishing the package
+COPY trubudget-pino-logger trubudget-pino-logger
+COPY server.js server.js
+COPY package*.json ./
 
 RUN npm ci
 
-# Use node user to execute the application. Node user is not in su list
 USER node
 
 CMD npm start
