--- logging-service/docker-compose.yml
@@ -1,14 +1,0 @@
-version: "3.2"
-services:
-  logging-service:
-    build:
-      context: .
-    environment:
-      LOGGER_PORT: ${LOGGER_PORT}
-      API_HOST: ${API_HOST}
-      API_PORT: ${API_PORT}
-      LOG_LEVEL: ${LOG_LEVEL}
-      LOGGING_SERVICE_CACHE_DURATION: ${LOGGING_SERVICE_CACHE_DURATION}
-      NODE_ENV: ${LOGGING_SERVICE_NODE_ENV}
-    ports:
-      - "${LOGGER_PORT}:${LOGGER_PORT}"
