+++ logging-service/package.json
@@ -1,8 +1,8 @@
 {
   "name": "trubudget-logging-service",
-  "version": "1.1.6",
-  "description": "This is module exports function that creates a pino logger instance with a commen configuration for the trubudget project",
-  "main": "index.js",
+  "version": "2.1.0",
+  "description": "This service is part of TruBudget. It collects logs of users using TruBudget's frontend",
+  "main": "server.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/openkfw/TruBudget.git"
@@ -12,11 +12,11 @@
   "bugs": {
     "url": "https://github.com/openkfw/TruBudget/issues"
   },
-  "homepage": "https://github.com/openkfw/TruBudget#readme",
+  "homepage": "https://github.com/openkfw/TruBudget/logging-service",
   "scripts": {
-    "start": "node ./serve.js",
+    "start": "node ./server.js",
     "audit": "better-npm-audit audit",
-    "dev": "nodemon ./serve.js"
+    "watch": "nodemon ./server.js"
   },
   "devDependencies": {
     "@types/node": "^16.9.6",
@@ -27,8 +27,6 @@
     "better-npm-audit": "^3.4.0",
     "fastify": "^3.22.1",
     "fastify-cors": "^6.0.2",
-    "pino": "^6.13.2",
-    "pino-http": "^5.8.0",
-    "pino-pretty": "^7.0.1"
+    "trubudget-pino-logger": "file:trubudget-pino-logger"
   }
 }
