+++ logging-service/trubudget-pino-logger/.npmignore
@@ -1,0 +1,1 @@
+node_modules
